feat: migrate router to the api/v1/router.py file
