+++ backend/apps/notifications/src/main.py
@@ -1,20 +1,17 @@
+import asyncio
+
 from logging import config as logging_config
 
 import structlog
-from faststream import Depends
+
 from faststream import FastStream
-from faststream.kafka.annotations import KafkaMessage
 from faststream.rabbit import RabbitBroker
 
 from src.common import brokers
-from src.emails import dependencies as emails_deps
-from src.emails.services import (
-    EmailNotificationService,
-    email_service_factory,
-    EmailServiceSchemaType,
-)
+from src.emails.api.v1.router import router as email_router
 from src.settings.app import get_app_settings
 from src.settings.logging import configure_logger
+
 
 settings = get_app_settings()
 
@@ -23,9 +20,10 @@ configure_logger(enable_async_logger=True)
 
 logger = structlog.get_logger()
 
+broker = brokers.get_kafka_broker()
 
 app = FastStream(
-    broker=brokers.get_kafka_broker(),
+    broker=broker,
     title=settings.service.name,
     description=settings.service.description,
     version=settings.service.version,
@@ -55,59 +53,8 @@ async def on_shutdown() -> None:
         await brokers.broker_rabbit.close()
 
 
-@app.broker.subscriber(
-    settings.notification.email_welcome_topic_name,
-    group_id=settings.notification.email_group_id,
-    batch=True,
-    auto_commit=False,
-    retry=5,
-)
-async def email_welcome_event_handler(
-    message_payload: list[emails_deps.WelcomeEventMessage],
-    msg_context: KafkaMessage,
-    service: EmailNotificationService = Depends(
-        email_service_factory(EmailServiceSchemaType.welcome)
-    ),
-) -> None:
-    await logger.info(
-        "Message has been consumed from Kafka",
-        topic_name=settings.notification.email_welcome_topic_name,
-        group_id=settings.notification.email_group_id,
-        message_id=msg_context.message_id,
-        message_payload=message_payload,
-    )
-
-    await service.handle_events(
-        event_messages=message_payload,
-        queue_name=settings.notification.email_welcome_queue_name,
-        msg_context=msg_context,
-    )
+broker.include_router(email_router)
 
 
-@app.broker.subscriber(
-    settings.notification.email_weekly_update_topic_name,
-    group_id=settings.notification.email_group_id,
-    batch=True,
-    auto_commit=False,
-    retry=5,
-)
-async def email_weekly_update_event_handler(
-    message_payload: list[emails_deps.WeeklyUpdateMessage],
-    msg_context: KafkaMessage,
-    service: EmailNotificationService = Depends(
-        email_service_factory(EmailServiceSchemaType.weekly_update)
-    ),
-) -> None:
-    await logger.info(
-        "Message has been consumed from Kafka",
-        topic_name=settings.notification.email_weekly_update_topic_name,
-        group_id=settings.notification.email_group_id,
-        message_id=msg_context.message_id,
-        message_payload=message_payload,
-    )
-
-    await service.handle_events(
-        event_messages=message_payload,
-        queue_name=settings.notification.email_weekly_update_queue_name,
-        msg_context=msg_context,
-    )
+if __name__ == "__main__":
+    asyncio.run(app.run())
